+++ GoodwingTimetabler/myTests/tests.py
@@ -7,8 +7,10 @@ def generateScheduleUsingCSP():
     print("generateScheduleUsingCSP() running...\n")
 
     # Create the university
-    my_univ = generateUniv("ESILV", dt.date(2025, 1, 6), 7, time_ranges)
-
+    #my_univ = generateUniv("ESILV", dt.date(2025, 1, 6), 7, time_ranges)
+    my_univ = generateUniv2()
+    print("Univ generated successfully : ", my_univ)
+    print("Starting solving the CSP...")
     # Instantiate and solve the CSP
     scheduler = CSP(my_univ)
 
@@ -19,21 +21,11 @@ def generateScheduleUsingCSP():
 def outputSchedulesFromCSP(csp_solver: CSP):
     group_courses = {}
 
-    print("\nDEBUG: Initial courses from CSP:")
-    for course in csp_solver.generated_courses:
-        print(f"Subject: {course.subject.name}, Day: {course.timeslot.day}, Time: {course.timeslot.start}-{course.timeslot.end}")
-
     # Organize courses by group
     for course in csp_solver.generated_courses:
         if course.group.name not in group_courses:
             group_courses[course.group.name] = []
         group_courses[course.group.name].append(course)
-
-    print("\nDEBUG: Courses after grouping:")
-    for group_name, courses in group_courses.items():
-        print(f"\nGroup: {group_name}")
-        for course in courses:
-            print(f"Subject: {course.subject.name}, Day: {course.timeslot.day}, Time: {course.timeslot.start}-{course.timeslot.end}")
 
     # Output each group's schedule
     for group_name, courses in group_courses.items():