+++ GoodwingTimetabler/util/__init__.py
@@ -1,2 +1,3 @@
 from .helloworld import hello_world
 from .yamlManager import *
+from .excelManager import *